feat: update application structure and components for legal chatbot, including new forms and layout adjustments
