+++ constants/app.ts
@@ -1,0 +1,11 @@
+import app from "../package.json";
+
+// MARK: - App Constants
+export const APP_NAME = app.name;
+export const APP_TAGLINE = app.tagline;
+export const APP_VERSION = app.version;
+
+// MARK: - Response statuses
+export const SUCCESS_STATUS = "success";
+export const FAILED_STATUS = "failed";
+export const INVALID_DATA_STATUS = "invalid_data";
