+++ lib/hooks/useCountryStateCitySelection.ts
@@ -2,34 +2,29 @@ import { useLookupStore } from "@/lib/store/lookupStore";
 import { useEffect, useState } from "react";
 
 export const useCountryStateCitySelection = () => {
-  const [countryId, setCountryId] = useState<string | null>(null);
-  const [deptoStateId, setDeptoStateId] = useState<string | null>(null);
+  const [countryId, setCountryId] = useState<string>();
+  const [deptoStateId, setDeptoStateId] = useState<string>();
   const lookupStore = useLookupStore();
 
+  // biome-ignore lint/correctness/useExhaustiveDependencies: only run on mount once, dependencies are not needed
   useEffect(() => {
-    if (lookupStore.countries.length === 0) {
-      lookupStore.fetchCountries();
-    }
-  }, [lookupStore, lookupStore.countries.length]);
+    lookupStore.fetchCountries();
+  }, []);
 
-  useEffect(() => {
+  const onCountryChange = (countryId: string) => {
+    // value is now the country ID directly
+    setCountryId(countryId);
     lookupStore.cleanDeptoStates();
     lookupStore.cleanCityMunicipalities();
-  }, [lookupStore]);
-
-  const onCountryChange = (value: string) => {
-    // value is now the country ID directly
-    const countryId = value;
-    setCountryId(countryId);
     if (countryId) {
       lookupStore.fetchDeptoStates(countryId);
     }
   };
 
-  const onStateChange = (value: string) => {
+  const onStateChange = (deptoStateId: string) => {
     // value is now the state ID directly
-    const deptoStateId = value;
     setDeptoStateId(deptoStateId);
+    lookupStore.cleanCityMunicipalities();
     if (countryId) {
       lookupStore.fetchCityMunicipalities(countryId);
     }
